requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cytoolz==0.12.0
ecdsa==0.18.0
eth-hash==0.3.3
eth-keys==0.4.0
eth-typing==3.1.0
eth-utils==2.0.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
more-itertools==8.13.0
multidict==6.0.2
nextcord==2.0.0
py-bip39-bindings==0.1.9
py-ed25519-bindings==1.0.1
py-sr25519-bindings==0.1.4
pycparser==2.21
pycryptodome==3.15.0
PyNaCl==1.5.0
python-dotenv==0.20.0
requests==2.28.1
scalecodec==1.0.38
six==1.16.0
SQLAlchemy==1.4.39
substrate-interface==1.2.7
toolz==0.12.0
urllib3==1.26.10
websocket-client==1.3.3
xxhash==2.0.2
yarl==1.7.2
